Refactored UserController and Created AdminController. Also, added user update for admin purposes.
