+++ src/main/java/com/project2/Project2/controller/UserController.java
@@ -4,12 +4,10 @@ import com.project2.Project2.model.User;
 import com.project2.Project2.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 
 @RestController
 @RequestMapping("/api/users")
@@ -22,13 +20,6 @@ public class UserController {
     public ResponseEntity<User> findUserById(@PathVariable String id) {
         Optional<User> user = userService.getUserById(id);
         return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-    }
-
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @PostMapping
-    public ResponseEntity<User> saveUser(@RequestBody User user) {
-        Optional<User> createdUser = Optional.ofNullable(userService.saveUser(user));
-        return createdUser.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PutMapping("/{id}")
@@ -60,33 +51,4 @@ public class UserController {
         Optional<List<String>> wishlist = userService.getUserWishlist(id);
         return wishlist.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
-
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @PutMapping("/{id}/roles")
-    public ResponseEntity<User> assignRolesToUser(@PathVariable String id, @RequestBody String roles) {
-        Optional<User> updatedUser = userService.assignRole(id, roles);
-        return updatedUser.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-    }
-
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @DeleteMapping("/{id}/roles")
-    public ResponseEntity<User> removeRolesFromUser(@PathVariable String id, @RequestBody String roles) {
-        Optional<User> updatedUser = userService.removeRole(id, roles);
-        return updatedUser.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-    }
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @GetMapping
-    public ResponseEntity<List<User>> getAllUsers() {
-        List<User> users = userService.getAllUsers();
-        return ResponseEntity.ok(users);
-    }
-
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @GetMapping("/{id}/roles")
-    public ResponseEntity<Set<String>> getUserRoles(@PathVariable String id) {
-        Optional<Set<String>> roles = userService.getUserRoles(id);
-        return roles.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
-    }
 }
-
-
